Backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cross-web==0.4.1
cryptography==46.0.5
ecdsa==0.19.1
fastapi==0.129.0
graphql-core==3.2.7
greenlet==3.3.1
h11==0.16.0
idna==3.11
packaging==26.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
strawberry-graphql==0.296.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
